rename and clean julearn.estimators to julearn.models
